Fix a few bugs (in non-debug config), clean up warnings
In several places, we called functions and relied on their
side-effects, within an assert() call. This means that when
building for non-debug configurations, the side-effect
expressions did not run, and so they never computed the
values that we needed (stuff we needed even beyond the
assert() call). This fixes them by adding an assert_or_ignore()
function, an ignore() function, and by moving some repeated code
to a function in codegen_mmir.cpp.

This also fixes several warnings. GCC and Clang do control-flow
analysis slightly differently. In some situations, GCC thinks that
a particular point in control-flow is reachable, while Clang thinks
that the same point is not reachable. This causes the analysis of
whether a function always returns a value to fail (erroneously).
This commit fixes this by removing a few "break;" statements that
were confusing GCC.